+++ src/decorators/Arg.ts
@@ -1,9 +1,18 @@
-import { ReturnTypeFunc, DecoratorTypeOptions, DescriptionOptions, ValidateOptions } from "./types";
+import {
+  ReturnTypeFunc,
+  DecoratorTypeOptions,
+  DescriptionOptions,
+  ValidateOptions,
+  DeprecationOptions,
+} from "./types";
 import { getMetadataStorage } from "../metadata/getMetadataStorage";
 import { getParamInfo } from "../helpers/params";
 import { getTypeDecoratorParams } from "../helpers/decorators";
 
-export type ArgOptions = DecoratorTypeOptions & DescriptionOptions & ValidateOptions;
+export type ArgOptions = DecoratorTypeOptions &
+  DescriptionOptions &
+  ValidateOptions &
+  DeprecationOptions;
 
 export function Arg(name: string, options?: ArgOptions): ParameterDecorator;
 export function Arg(
@@ -25,6 +34,7 @@ export function Arg(
       kind: "arg",
       name,
       description: options.description,
+      deprecationReason: options.deprecationReason,
       ...getParamInfo({
         prototype,
         propertyKey,